Initial prompt
Ok, now let's create an LWC. I want an LWC that will go on the Contact record page. I should display the name, title, company name (and hyperlink to the Account), and the avatar picture, a URL stored in Avatar__c on the Contact record. This should fit in style with the Salesforce Lightning Design System, and be similar to a LinkedIn profile card, as the function will be the same, to show a graphical, immediate representation of the Contact.