Cursor Edit
Ah, you're right. In LWC templates, we can't use optional chaining. We need to use the getter we created in the JavaScript file. Let me fix the HTML template:

I've updated the template to:

 - Use the getters we defined in the JavaScript file (`name`, `title`, `avatar`, `account`)
 - Remove the optional chaining operator
 - Access the account properties through the `account` getter

This should now work correctly. The template will use the values from our getters, which handle the data access and null checking in the JavaScript file.